+++ src/si/extended/operations_generic.rs
@@ -754,7 +754,8 @@ where
         QuantityGeneric::<L, M, T, I, TH, N, LUM, D_TH, A, SA, INFO, Storage>(x * unit.0)
     }
 
-    /// Returns dimensional formula of a quantity
+    /// Returns dimensional formula array of a quantity
+    /// [L, M, T, I, TH, N, LUM, D_TH, A, SA, INFO]
     pub const fn dim(&self) -> [i64; 11] {
         [L, M, T, I, TH, N, LUM, D_TH, A, SA, INFO]
     }
@@ -768,6 +769,59 @@ where
         unit: QuantityGeneric<L, M, T, I, TH, N, LUM, D_TH, A, SA, INFO, Storage>,
     ) -> Storage {
         self.0 / unit.0
+    }
+
+    #[cfg(feature = "std")]
+    /// Return units-of-measure string  
+    /// ```rust
+    /// #![feature(generic_const_exprs)]
+    /// use dimensional_quantity::si::extended::f64::quantities::LuminousEfficacy;
+    /// let l = LuminousEfficacy::new(1.0);
+    /// assert_eq!(l.si_uom_str(), "m⁻²⋅kg⁻¹⋅s³⋅cd⋅sr");
+    /// assert_eq!(l.powi::<5>().si_uom_str(),"m⁻¹⁰⋅kg⁻⁵⋅s¹⁵⋅cd⁵⋅sr⁵");
+    /// ```
+    pub fn si_uom_str(&self) -> String {
+        use crate::format::dim_to_string;
+        use std::fmt::Write;
+
+        let mut output = String::new();
+        if let Some(dim) = dim_to_string(L) {
+            let _ = write!(output, "m{}⋅", dim);
+        }
+        if let Some(dim) = dim_to_string(M) {
+            let _ = write!(output, "kg{}⋅", dim);
+        }
+        if let Some(dim) = dim_to_string(T) {
+            let _ = write!(output, "s{}⋅", dim);
+        }
+        if let Some(dim) = dim_to_string(I) {
+            let _ = write!(output, "A{}⋅", dim);
+        }
+        if let Some(dim) = dim_to_string(TH) {
+            let _ = write!(output, "K{}⋅", dim);
+        }
+        if let Some(dim) = dim_to_string(D_TH) {
+            let _ = write!(output, "ΔK{}⋅", dim);
+        }
+        if let Some(dim) = dim_to_string(N) {
+            let _ = write!(output, "mol{}⋅", dim);
+        }
+        if let Some(dim) = dim_to_string(LUM) {
+            let _ = write!(output, "cd{}⋅", dim);
+        }
+        if let Some(dim) = dim_to_string(A) {
+            let _ = write!(output, "rad{}⋅", dim);
+        }
+        if let Some(dim) = dim_to_string(SA) {
+            let _ = write!(output, "sr{}⋅", dim);
+        }
+        if let Some(dim) = dim_to_string(INFO) {
+            let _ = write!(output, "bit{}⋅", dim);
+        }
+
+        let mut output = output.chars();
+        output.next_back();
+        output.collect()
     }
 }
 
@@ -958,7 +1012,7 @@ where
     ///
     ///
     /// ```
-    /// 
+    ///
     #[cfg(feature = "std")]
     pub fn cbrt(
         &self,